Restructure the queue service as a resources/js module
Follows the shared module pattern (like listbox): the service moves to
resources/js/modules/queue/queue.ts with its types and test alongside,
and a new index.ts assigns Craft.QueueService for legacy/plugin callers.
Both entrypoints load the shim.

CP.js keeps importing the service directly rather than using the global,
because the webpack bundle instantiates it while still evaluating -
before Vite's module scripts have assigned Craft.QueueService.
