+++ resources/js/common/types/keys.ts
@@ -1,5 +1,5 @@
 import type {InjectionKey} from 'vue';
-import type {QueueService} from '@/common/services/Queue';
+import type {QueueService} from '@/modules/queue/queue';
 import type {AxiosInstance} from 'axios';
 import type {ConfigService} from '@craftcms/ui/services/Config';
