+++ resources/js/legacy.ts
@@ -13,6 +13,7 @@ import './modules/listbox/index';
 import './modules/field-layout-designer/index';
 import './modules/sortable-checkbox-select/index';
 import './modules/editable-table/index';
+import './modules/queue/index';
 
 const {default: Cp} = await import('./bootstrap/cp.js');
