+++ resources/js/modules/queue/index.ts
@@ -1,0 +1,17 @@
+import {QueueService} from './queue';
+
+/**
+ * Legacy and plugin code reaches the queue through the `Craft` global rather
+ * than an ES module import, so expose the service class there
+ * (`Craft.QueueService.getInstance()`).
+ *
+ * Note that the legacy webpack bundle's `CP.js` cannot use this global — it
+ * needs the service while its own bundle is still evaluating, before Vite's
+ * module scripts have run — so it imports `./queue` directly and bundles its
+ * own copy.
+ */
+const craft = (window as any).Craft ?? ((window as any).Craft = {});
+craft.QueueService = QueueService;
+
+export {QueueService};
+export * from './types';
