+++ resources/js/modules/queue/types.ts
