+++ resources/js/modules/utilities/components/queue-manager/QueueManagerToolbar.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-  import {type JobInfo, JobStatus} from '@/common/types/queue';
+  import {type JobInfo, JobStatus} from '@/modules/queue/types';
   import {t} from '@craftcms/ui/utilities/translate';
   import {computed, ref, watch} from 'vue';
   import {useFlashMessages} from '@/common/composables/useFlashMessages';
